Задача U12. Alerts (#13)
* Задача U0. Настройка окружения

* Задача U1. 1-3

* Задача U1. 5

* Задача U1. 4

* Задача U1. 4

* Задача U1. fix

* Задача U1. fix

* Задача U1. fix

* Задача U2. Allure

* Задача U2. fix

* Задача U2. fix

* Задача U2. fix

* Задача U3. DataProvider

* Задача U3. DataProvider

* Задача U4. Screenshots

* Задача U4. Screenshots

* Задача U4. Screenshots

* Задача U5. Cookies

* Задача U6. JavaScriptExecutor

* Задача U7. Параллельный запуск тестов

* Задача U7. Параллельный запуск тестов

* Задача U7. Параллельный запуск тестов. Исправление

* Задача U7. Параллельный запуск тестов. Исправление

* Задача U8. Перезапуск упавших тестов

* Задача U9. Браузеры

* Задача U9. Браузеры

* Задача U10. Drag n Drop (IFrame)

* Задача U11. Tabs

* Задача U11. Tabs

* Задача U12. Alerts

+++ src/test/java/utils/CookieUtils.java
@@ -20,14 +20,28 @@ public class CookieUtils {
     }
 
     public static void loadCookies(WebDriver driver) {
-        try (ObjectInputStream ois = new ObjectInputStream(new FileInputStream(COOKIE_FILE))) {
-            driver.manage().deleteAllCookies();
-            Set<Cookie> cookies = (Set<Cookie>) ois.readObject();
-            for (Cookie cookie : cookies) {
-                driver.manage().addCookie(cookie);
-            }
-        } catch (Exception e) {
+        File file = new File(COOKIE_FILE);
+        if (!file.exists()) {
             System.out.println("Cookie file not found — авторизуемся заново");
+            return;
+        }
+
+        try (ObjectInputStream ois = new ObjectInputStream(new FileInputStream(file))) {
+            Set<Cookie> cookies = (Set<Cookie>) ois.readObject();
+            if (cookies == null || cookies.isEmpty()) return;
+
+            driver.manage().deleteAllCookies();
+
+            for (Cookie cookie : cookies) {
+                try {
+                    driver.manage().addCookie(cookie);
+                } catch (Exception e) {
+                    System.out.println("Не удалось добавить куку " + cookie.getName() + ": " + e.getMessage());
+                }
+            }
+
+        } catch (Exception e) {
+            System.out.println("Ошибка при загрузке cookies: " + e.getMessage());
         }
     }
